# api/openapi-lock.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "org-quicko-cliq",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://{{base_url}}"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
  "paths": {
...
    "/programs/{program_id}/promoters/{promoter_id}/members/{member_id}": {
      "patch": {
        "tags": [
          "Promoter"
        ],
        "summary": "Update role",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "role": ""
                }
              },
              "examples": {
                "success": {
                  "value": {
                    "role": ""
                  }
                },
                "failure- unauthorized": {
                  "value": {
                    "role": ""
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          },
          {
            "name": "promoter_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{promoter_id}}"
          },
          {
            "name": "member_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{member_id}}"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully changed user role.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"error\": \"You are unauthorized to perform this action.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Promoter"
        ],
        "summary": "Remove member",
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          },
          {
            "name": "promoter_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{promoter_id}}"
          },
          {
            "name": "member_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{member_id}}"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully removed member.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"error\": \"You are unauthorized to perform this action.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      }
    },
...
    "/members/signup": {
      "post": {
        "tags": [
          "Member"
        ],
        "summary": "Member sign up",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "@entity": {
                    "default": "member"
                  },
                  "member_id": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  },
                  "first_name": {
                    "type": "string"
                  },
                  "last_name": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "member"
                    ]
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              },
              "examples": {
                "success": {
                  "value": {
                    "email": "leonardo@mail.com",
                    "password": "2j3n2o30j29j",
                    "first_name": "Leonardo",
                    "last_name": "Da Vinci"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully signed up member.\",\r\n    \"data\": {\r\n        \"member_id\": \"wevn1i23yd13hcoiu3bnvkn\"\r\n    }\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {
                  "201-Example-1": {
                    "value": {
                      "email": "",
                      "password": "",
                      "first_name": "",
                      "last_name": ""
                    },
                    "summary": "Example 1 for status 201"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/members/login": {
      "post": {
        "tags": [
          "Member"
        ],
        "summary": "Member log in",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "@entity": {
                    "default": "member"
                  },
                  "member_id": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  },
                  "first_name": {
                    "type": "string"
                  },
                  "last_name": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "member"
                    ]
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              },
              "examples": {
                "success": {
                  "value": {
                    "email": "",
                    "password": ""
                  }
                },
                "failure- incorrect credentials": {
                  "value": {
                    "email": "",
                    "password": ""
                  }
                }
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"error\": \"Incorrect credentials. Please try again.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {
                  "401-Example-1": {
                    "value": {
                      "email": "luke@mail.com",
                      "password": "notnice"
                    },
                    "summary": "Example 1 for status 401"
                  }
                }
              }
            }
          },
          "undefined": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"data\": {\r\n        \"token\": \"2kj3nkj3nco9h9%^82ewh\"\r\n    },\r\n    \"message\": \"Successfully logged in member.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {
                  "undefined-Example-1": {
                    "value": {
                      "email": "luke@mail.com",
                      "password": "notnice"
                    },
                    "summary": "Example 1 for status undefined"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/members/{member_id}": {
      "get": {
        "tags": [
          "Member"
        ],
        "summary": "Get member",
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{member_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully got member.\",\r\n    \"data\": {\r\n        \"member\": {\r\n            \"email\": \"leonardo@mail.com\",\r\n            \"password\": \"2j3n2o30j29j\",\r\n            \"first_name\": \"Leonardo\",\r\n            \"last_name\": \"Da Vinci\"\r\n        }\r\n    }\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Member"
        ],
        "summary": "Update member info",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "@entity": {
                    "default": "member"
                  },
                  "member_id": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  },
                  "first_name": {
                    "type": "string"
                  },
                  "last_name": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "member"
                    ]
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              },
              "examples": {
                "success": {
                  "value": {
                    "email": "",
                    "password": "",
                    "first_name": "",
                    "last_name": ""
                  }
                },
                "failure- unauthorized": {
                  "value": {
                    "email": "",
                    "password": "",
                    "first_name": "",
                    "last_name": ""
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{member_id}}"
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"error\": \"You are unauthorized to perform this action.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          },
          "undefined": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully updated member information.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Member"
        ],
        "summary": "Delete member",
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{member_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
...
  }
}